requirements.txt
et-xmlfile==1.1.0
openpyxl==3.1.2
python-decouple==3.8
python-dotenv==1.0.1
drf-yasg==1.21.7
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-filter==23.5
django-resized==1.0.2
djangorestframework==3.14.0